requirements.txt
fastapi
uvicorn
pandas
requests
feedparser
nltk
dash
plotly
dash-bootstrap-components
wordcloud
psycopg2-binary